Correct simulation. Cleaned up everything.
