+++ .envrc
@@ -6,6 +6,7 @@ export PYTHONPATH=$PWD/submodules/robupy
 export PYTHONPATH=$PYTHONPATH:$PWD/submodules/ruspy
 export PYTHONPATH=$PYTHONPATH:$PWD/submodules/zurcher-data
 export PYTHONPATH=$PYTHONPATH:$PWD/python
+export PYTHONPATH=$PYTHONPATH:$PWD/python/figures_application_scripts
 
 # We need to be able execute our auxilary scripts
 export PATH=$PWD/bin:$PATH
