+++ .gitignore
@@ -6,12 +6,11 @@ __pycache__/
 */jupyter/migrated*
 *.pbs
 *.png
-fixp_results_5000_50_400_2223.pkl
-fixp_results_5000_50_400_4292.pkl
+fixp_results_2223.pkl
+fixp_results_4292.pkl
 sim_results/
 val_results/
 validation_results/
-validation_results_revisited.zip
 demonstration.pkl
