--- bin/execute-notebooks
@@ -1,1 +1,0 @@
-../python/execute_notebooks.py
