--- bin/update-repository
@@ -1,1 +1,0 @@
-../python/update_repository.py
