+++ configurations/ipython/profile_default/startup/profile_default/startup/start.py
