--- python/general_specification.json
@@ -1,38 +1,0 @@
-{
-  "num_points": 100,
-  "max_omega": 0.99,
-  "num_workers": 100,
-  "num_states": 50,
-  "sample_size": 4292,
-  "scale": 78,
-  "beta": 0.9999,
-  "threshold": 1e-05,
-  "params": [
-    50,
-    400
-  ],
-  "seed": 123,
-  "buses": 1000,
-  "periods": 100000,
-  "sim_keys": [
-    0.0,
-    0.5,
-    0.95
-  ],
-  "density_strategy": 0.5,
-  "density_runs": 10000,
-  "runs_strategies_validation": 10000,
-  "strategies_validation": [
-    0.1,
-    0.2,
-    0.3,
-    0.4,
-    0.6,
-    0.7,
-    0.8,
-    0.9
-  ],
-  "trans_probs": "../../pre_processed_data/parameters/rust_trans_probs.txt",
-  "raw_trans": "../../pre_processed_data/parameters/rust_trans_raw.txt",
-  "raw_cov": "../../pre_processed_data/parameters/rust_cov_raw.txt"
-}
