--- python/solution/specification.json
@@ -1,15 +1,0 @@
-{
-  "num_points": 100,
-  "max_omega": 0.99,
-  "num_states": 50,
-  "sample_size": 4292,
-  "scale": 78,
-  "beta": 0.9999,
-  "threshold": 1e-05,
-  "num_workers": 100,
-  "trans_probs": "../../pre_processed_data/parameters/rust_trans_probs.txt",
-  "params": [
-    50,
-    400
-  ]
-}
