--- python/validation/specification.json
@@ -1,27 +1,0 @@
-{
-  "num_workers": 100,
-  "beta": 0.9999,
-  "buses": 1000,
-  "periods": 100000,
-  "params": [
-    50,
-    400
-  ],
-  "seed": 123,
-  "density_strategy": 0.5,
-  "density_runs": 10000,
-  "runs_strategies_validation": 10000,
-  "strategies_validation": [
-    0.1,
-    0.2,
-    0.3,
-    0.4,
-    0.6,
-    0.7,
-    0.8,
-    0.9
-  ],
-  "raw_trans": "../../pre_processed_data/parameters/rust_trans_raw.txt",
-  "raw_cov": "../../pre_processed_data/parameters/rust_cov_raw.txt",
-  "policy_dict": "../../pre_processed_data/fixp_results_5000_50_400_4292.pkl"
-}
